% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/david2014.R
\docType{data}
\name{david2014}
\alias{david2014}
\title{David2014}
\format{The data set in \code{\link{phyloseq-class}} format}
\usage{
david2014<-david_2014
}
\value{
Loads the data set in R.
}
\description{
Consists high-throughput profiling data collected from gut microbiota samples collected from two individuals and Saliva samples collected form one of them.
Gut microbiota samples from Subject A collected between  days 0 and 364 of the study and saliva microbiota samples between days 26 and 364.
In case of Subject B, gut microbiota samples between study days 0 and 252 were collected.
For details, see \url{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2014-15-7-r89}.
}
\references{
David et al. (2014)
\url{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2014-15-7-r89}
}
\author{
Binu Mathew \email{binmat@utu.fi}
}
\keyword{data}
